[A1-fix] fix(build): move buildGuideRoleFlipWrite out of the "use server" module
`bun run build` failed: "Server Actions must be async functions". A "use server"
module may only export async functions, and A1 exported a SYNC pure helper from
actions.ts to make the decision testable -- which is exactly what the plan told it
to do ("in actions.ts (exported for tests)"). The plan was wrong.

Nothing else caught this. Not typecheck, not lint, not the 1279 tests, not the
pre-commit hook (which runs typecheck + lint + tests, but NOT the build). Only
`next build` knows this rule.

The repo already had the right home for it: src/data/admin-users/guards.ts, whose
own docblock says "pure functions (no server imports) so they can be shared by
server actions AND client islands, and unit-tested in isolation" -- it is where
evaluateRoleChange and evaluateStatusChange already live. Behaviour is unchanged;
the function just sits where the convention says it belongs.

The test now imports it from @/data/admin-users and says why, so the next person
does not move it back.

proof: bun run build -> PASS (was: FAIL). typecheck clean; lint 0 errors;
bun run test:run -> 234 files / 1279 tests pass.
