+++ src/app/(protected)/admin/users/actions.ts
@@ -3,6 +3,7 @@
 import { revalidatePath } from "next/cache";
 
 import {
+  buildGuideRoleFlipWrite,
   bulkActionInputSchema,
   canHardDeleteUser,
   evaluateRoleChange,
@@ -17,7 +18,6 @@ import {
   type AdminActionResult,
   type BulkActionResult,
 } from "@/data/admin-users";
-import { COPY } from "@/lib/copy";
 import { createSupabaseServerClient } from "@/lib/supabase/server";
 import {
   countOtherActiveAdmins,
@@ -53,46 +53,6 @@ function ok(message: string): AdminActionResult {
 }
 
 type AdminClient = Awaited<ReturnType<typeof requireAdminSession>>["adminClient"];
-
-/**
- * Decide what a flip-to-guide may write to `guide_profiles`.
- *
- * The bug (item 13, msgs 544–593): this path used to upsert `display_name`,
- * `verification_status` and `is_available` unconditionally, so every role flip
- * overwrote an existing guide's public identity with their email local-part,
- * reset them to `draft` and hid them from `/guides`. Admins could not see it —
- * RLS shows them `profiles.full_name` while anonymous visitors read the
- * clobbered `guide_profiles.display_name`. Re-approving never healed it.
- *
- * An existing profile is therefore never rewritten. A new one falls back to the
- * full name and NEVER to the email local-part — that leak is the same bug.
- */
-export function buildGuideRoleFlipWrite(
-  existing: { user_id: string; verification_status: string } | null,
-  input: { targetId: string; targetFullName: string | null },
-):
-  | { kind: "insert"; row: Record<string, unknown> }
-  | { kind: "restore-visibility" }
-  | { kind: "noop" } {
-  if (existing) {
-    // D-A1a: the demote branch below is what forced is_available=false, so
-    // re-promoting an approved guide undoes exactly that and nothing else.
-    // A guide who wants to stay hidden can re-toggle it themselves.
-    return existing.verification_status === "approved"
-      ? { kind: "restore-visibility" }
-      : { kind: "noop" };
-  }
-
-  return {
-    kind: "insert",
-    row: {
-      user_id: input.targetId,
-      display_name: input.targetFullName?.trim() || COPY.guide,
-      verification_status: "draft",
-      is_available: false,
-    },
-  };
-}
 
 async function syncRoleSideEffects(input: {
   adminClient: AdminClient;
